refactor(app): extract AppOverlays from App.tsx (#316 phase 4)
Moves all 19 screen-independent overlays into app/AppOverlays.tsx, the first
consumer of useApp(): event-card reveal, the three campaign confirm modals,
relic spin, integrity warning, achievement toasts, the service-worker update
prompt, cloud-sync prompt, gift claim, wins-milestone celebration, streak
broken, time capsule, deck selector, exotic drop, quest completes, chronicle
completes, weekly reward and the daily-login modal.

Every block is a verbatim move, including the daily-login onClose
reward-granting branch.

Ordering note: the event-card reveal previously rendered before the screen
branches and now renders after them with the rest. Every overlay involved
carries an explicit z-index in styles.css (400 through 9999) and screens set
none at the root, so stacking is unchanged — will confirm on the dev server
at the routing checkpoint.

AppContextValue grows to cover what the overlays read. Also exports SyncPrompt
from useCloudSync so the context can name the type.

Pre-existing dead code left alone, flagged rather than fixed here:
handleGameOverPrimary, showFingerSmash and one destructured binding are unused
on main too (verified by re-running tsc --noUnusedLocals against the base
commit). handleGameOverPrimary is a 57-line handler nothing calls and is worth
its own issue.

App.tsx: 3,734 -> 3,402 lines. Build green, chunk boundaries unchanged,
2107 tests passing.
